+++ doc/changelog.rst
@@ -12,6 +12,7 @@ Unreleased
 Core
 ~~~~
 
+- Fix refunds for discount lines
 - Fix restocking issue when refunding unshipped products
 - Make payments for `CustomPaymentProcessor` not paid by default
 - Fix shipping status for orders with refunds